correction methods and bug fixes

+++ src/main/resources/init_db.sql
@@ -3,6 +3,6 @@ USE books;
 CREATE TABLE books (
     id BIGINT NOT NULL AUTO_INCREMENT,
     title VARCHAR(255),
-    price INT,
+    price DECIMAL,
     PRIMARY KEY (id)
 );
